now sends request only if selected

+++ annuaire-web/src/main/webapp/resources/js/index.js
@@ -2,9 +2,14 @@ function processSelected(input) {
     var form = document.getElementById("contactsForm");
     var url = form.action;
 
-    if (input.id == 'deleteButton'){
-        form.action = url + "?method=delete";
-    } else if (input.id == 'emailButton'){
+    if (input.id == 'deleteButton') {
+        if (someSelected()) {
+            form.action = url + "?method=delete";
+        }
+        else {
+            return false;
+        }
+    } else if (input.id == 'emailButton') {
         form.action = url + "?method=email";
     }
